requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-JWT-Extended==4.4.4
Werkzeug==2.2.3
gunicorn==20.1.0
pytest==7.2.1